added additional error handling. Fixed errors with SL logic.

+++ src/main/java/com/company/rentalstorerestwebservicegroupproject/controller/CustomerViewModelController.java
@@ -11,6 +11,7 @@ import javax.validation.Valid;
 import javax.validation.constraints.Digits;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @RestController
 @Validated
@@ -22,6 +23,9 @@ public class CustomerViewModelController {
     @RequestMapping(value = "/customer", method = RequestMethod.GET)
     @ResponseStatus(value = HttpStatus.OK)
     public List<CustomerViewModel> getAllCustomerVMs() {
+        List<CustomerViewModel> customerViewModels = sl.findAllCustomerViewModels();
+        if (customerViewModels == null || customerViewModels.isEmpty())
+            throw new NoSuchElementException("No Customers found :-(");
         return sl.findAllCustomerViewModels();
     }
 
@@ -30,7 +34,6 @@ public class CustomerViewModelController {
     public String updateCustomers(@RequestBody @Valid List<CustomerViewModel> customerViewModels) {
         customerViewModels.stream().forEach(cvm -> sl.updateCustomerViewModel(cvm));
         return String.format("Successfully updated %s customers.", customerViewModels.size());
-
     }
 
     @RequestMapping(value = "/customer", method = RequestMethod.POST)
@@ -45,7 +48,7 @@ public class CustomerViewModelController {
     public List<CustomerViewModel> getCustomerById(@PathVariable @Digits(integer = 11, fraction = 0) Integer id) {
         CustomerViewModel cvm = sl.findCustomerViewModel(id);
         List<CustomerViewModel> cvmList = new ArrayList<>();
-        if (cvm == null) return cvmList;
+        if (cvm == null) throw new NoSuchElementException(String.format("Customer Id %s doesn't exist.", id));
         cvmList.add(cvm);
         return cvmList;
     }
@@ -53,6 +56,8 @@ public class CustomerViewModelController {
     @RequestMapping(value = "/customer/{id}", method = RequestMethod.DELETE)
     @ResponseStatus(value = HttpStatus.OK)
     public String deleteCustomerById(@PathVariable @Digits(integer = 11, fraction = 0) Integer id) {
+        CustomerViewModel cvm = sl.findCustomerViewModel(id);
+        if (cvm == null) throw new NoSuchElementException(String.format("Customer Id %s doesn't exist.", id));
         sl.deleteCustomerViewModel(id);
         return String.format("Successfully deleted Customer with id %s.", id);
     }
